+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 3: DOM Manipulation & Events
     - DOM Selectors For Multiple Elements
+    - Traversing The DOM
 - An example with DOM selectors for multiple elements was added.
+- An example with DOM traversing was added.
 
 ## August 17, 2020
 - Modern JavaScript From The Beginning (Udemy)